fix(protocols): avoid out-of-order table replies
sending table requests as fast as possible with libevent causes out of
order replies from RegattaCenter as it does not process the requests in
the same order or at least the replies are not sent in order of the
according requests. This commit makes the data protocol stateful, i.e.
it will not send new requests before any (!) full table reply will be
sent back to the gateway in response (MIC TV protocol ensures this
hopefully).
Additionally a fix in the buffer handling is given.