feat(protocols): resume activity after 2.5 s timeout
If the data connection is waiting for an reply but does not receives one
after 2.5 seconds, the state is reset to idle (which allows transmission
of further requests) after 2.5 seconds. This enabled recovery to an
operational state if a (malformed) request was issued that does not
return any answer at all.